LPS-49545 Moved LDAPFilterValidator related code to modules/portal/portal-ldap
