+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/validator/parser/LDAPFilterLexer.java
@@ -14,7 +14,7 @@
  * details.
  */
 
-package com.liferay.portal.parsers.ldap.parser;
+package com.liferay.portal.ldap.validator.parser;
 
 
 import org.antlr.runtime.*;
